requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
Werkzeug==2.3.7
pdfminer.six==20221105
gunicorn==21.2.0
